requirements.txt
# Local AI Generator for Aseprite - Python Dependencies
# Install with: pip install -r requirements.txt

# Core libraries
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
pillow>=9.5.0
numpy
openai>=1.0.0